refactor(cli): Purity C — engine uses EngineApiClient, delete CliStreamState
The seam flip left the CLI's AnthropicRuntimeClient in build_runtime even
though its inline render was already DEAD (set_spinner/set_output_writer
never called post-flip — rendering goes ObserverAdapter -> EngineEvent ->
EngineEventRenderer). Complete the cut: the engine now uses the pure
engine_core::EngineApiClient, and the old client + its rendering are gone.

- build_runtime / BuiltRuntime: ConversationRuntime<AnthropicRuntimeClient>
  -> <EngineApiClient>. EngineApiClient gains the two inherent methods the
  CLI reaches through api_client(): session_tracer() and
  fixed_request_overhead_tokens() (the budget-aware pre-send compaction on
  the engine path now works against the pure client).
- DELETE AnthropicRuntimeClient + CliStreamState + SpinnerRetryNotifier +
  their impls + the now-dead duplicate wire helpers (push_output_block/
  response_to_events/convert_messages/push_thinking_event/
  build_non_streaming_fallback_request/request_ends_with_tool_result/
  filter_tool_specs/prompt_cache_record_to_runtime_event/runtime_error_from_api).
  cli/api_client.rs is now just the finished-turn JSON collectors
  (final_assistant_text / collect_tool_uses|results|prompt_cache_events /
  max_tokens_for_model) that other paths still use.
- Friendly API errors: MOVE format_user_visible_api_error +
  format_context_window_blocked_error from cli/format.rs into the api crate
  (they only touch api::ApiError; use the existing truncate_body_snippet).
  EngineApiClient carries session_id and formats provider errors friendly
  in its stream/prefetch/mid-stream/fallback paths — restoring parity for
  network/auth/rate-limit display (context-window was already friendly via
  SessionEngine::run_turn -> context_overflow_user_message).

Verified: full workspace compiles; api unit tests; Windows mock
pty_core_conversation (streaming), pty_error_paths, output_format_contract
(JSON collectors) all green; fmt clean.
