+++ rust/crates/rusty-sudocode-cli/src/cli/api_client.rs
@@ -1,693 +1,16 @@
-use std::collections::VecDeque;
-use std::io::{self, Write};
+//! Finished-turn result collectors for the CLI's `--output-format json` /
+//! status paths.
+//!
+//! The provider client, the incremental wire→event translation, and the inline
+//! markdown/spinner rendering that used to live here have moved across the
+//! engine↔renderer seam: the pure client is now `engine_core::EngineApiClient`,
+//! the display half is `render_engine::EngineEventRenderer`, and friendly
+//! API-error formatting lives in the `api` crate
+//! (`api::format_user_visible_api_error`). What remains are small pure helpers
+//! that shape a finished [`runtime::TurnSummary`] into the JSON the one-shot /
+//! status paths emit — no I/O, no rendering.
 
-use api::{
-    resolve_startup_auth_source, AnthropicClient, AuthMode, AuthSource, CacheHints,
-    ContentBlockDelta, ImageSource, InputContentBlock, InputMessage, MessageRequest,
-    MessageResponse, OutputContentBlock, PromptCache, ProviderClient as ApiProviderClient,
-    StreamEvent as ApiStreamEvent, ToolChoice, ToolDefinition, ToolResultContentBlock,
-};
-use async_trait::async_trait;
-use futures::StreamExt;
-use runtime::{
-    ApiClient, ApiRequest, AssistantEvent, AssistantEventStream, ContentBlock, ConversationMessage,
-    MessageRole, PromptCacheEvent, RuntimeError, TokenUsage,
-};
-use telemetry::{SessionTracer, SudoclawLogSink};
-use tools::GlobalToolRegistry;
-
-use super::format::{format_tool_call_start, format_user_visible_api_error};
-use crate::render::{
-    query_terminal_width, MarkdownStreamState, ResponseGlyphState, SpinnerRef, TerminalRenderer,
-    BOLD, DIM, RESET,
-};
-use crate::repl_ui::OutputSender;
-use std::sync::Arc;
-
-use crate::{
-    AllowedToolSet, InternalPromptProgressReporter, RuntimeConfig, POST_TOOL_STALL_TIMEOUT,
-};
-
-/// Bridges HTTP retry notifications from the transport layer to the
-/// spinner's `set_retry` / resume methods, so the iocraft TurnRenderer
-/// can display retry status in the spinner overlay.
-struct SpinnerRetryNotifier {
-    spinner: SpinnerRef,
-}
-
-impl api::RetryNotifier for SpinnerRetryNotifier {
-    fn on_retry(&self, attempt: u32, max_retries: u32, reason: &str) {
-        self.spinner
-            .set_retry(attempt, max_retries, reason.to_string());
-    }
-
-    fn on_retry_end(&self) {
-        self.spinner.resume();
-    }
-}
-
-// NOTE: Despite the historical name `AnthropicRuntimeClient`, this struct
-// now holds an `ApiProviderClient` which dispatches to Anthropic, xAI,
-// OpenAI, or DashScope at construction time based on
-// `detect_provider_kind(&model)`. The struct name is kept to avoid
-// churning `BuiltRuntime` and every Deref/DerefMut site that references
-// it. See ROADMAP #29 for the provider-dispatch routing fix.
-pub(crate) struct AnthropicRuntimeClient {
-    pub(crate) runtime: tokio::runtime::Runtime,
-    pub(crate) client: ApiProviderClient,
-    pub(crate) session_id: String,
-    pub(crate) model: String,
-    pub(crate) enable_tools: bool,
-    pub(crate) emit_output: bool,
-    pub(crate) allowed_tools: Option<AllowedToolSet>,
-    pub(crate) tool_registry: GlobalToolRegistry,
-    pub(crate) progress_reporter: Option<InternalPromptProgressReporter>,
-    pub(crate) reasoning_effort: Option<String>,
-    /// Enable extended thinking for Anthropic models.
-    pub(crate) thinking_enabled: bool,
-    /// Shared spinner reference for pausing, thinking indicator, and
-    /// response-byte counting.
-    pub(crate) spinner: Option<SpinnerRef>,
-    /// Optional channel-backed writer that routes output through the iocraft
-    /// render loop instead of writing directly to stdout.
-    pub(crate) output_writer: Option<OutputSender>,
-}
-
-impl AnthropicRuntimeClient {
-    pub(crate) fn new(
-        session_id: &str,
-        config: &RuntimeConfig,
-        tool_registry: GlobalToolRegistry,
-    ) -> Result<Self, Box<dyn std::error::Error>> {
-        let sudocode_config = &config.sudocode_config;
-        let effective_mode = config.auth_mode;
-
-        let resolved = api::resolve_provider_from_config(
-            &config.model,
-            Some(effective_mode),
-            sudocode_config,
-        )?;
-        let mut client = ApiProviderClient::from_resolved(&resolved, Some(effective_mode))?
-            .with_prompt_cache(PromptCache::new(session_id));
-
-        // 默认启用日志追踪
-        let sink = Arc::new(SudoclawLogSink::new()?);
-        let tracer = SessionTracer::new(session_id, sink);
-        client = client.with_session_tracer(tracer);
-
-        Ok(Self {
-            runtime: tokio::runtime::Runtime::new()?,
-            client,
-            session_id: session_id.to_string(),
-            model: resolved.model_id.clone(),
-            enable_tools: config.enable_tools,
-            emit_output: config.emit_output,
-            allowed_tools: config.allowed_tools.clone(),
-            tool_registry,
-            progress_reporter: config.progress_reporter.clone(),
-            reasoning_effort: None,
-            thinking_enabled: true,
-            spinner: None,
-            output_writer: None,
-        })
-    }
-
-    pub(crate) fn set_spinner(&mut self, ref_: SpinnerRef) {
-        let notifier = Arc::new(SpinnerRetryNotifier {
-            spinner: ref_.clone(),
-        });
-        self.client.set_retry_notifier(notifier);
-        self.spinner = Some(ref_);
-    }
-
-    pub(crate) fn set_output_writer(&mut self, writer: OutputSender) {
-        self.output_writer = Some(writer);
-    }
-
-    /// Pause the spinner and clear its line before writing content.
-    fn pause_spinner(&self) {
-        if let Some(s) = &self.spinner {
-            s.pause();
-        }
-    }
-
-    /// Resume the spinner after content has been written.
-    fn resume_spinner(&self) {
-        if let Some(s) = &self.spinner {
-            s.resume();
-        }
-    }
-
-    pub(crate) fn set_reasoning_effort(&mut self, effort: Option<String>) {
-        self.reasoning_effort = effort;
-    }
-
-    pub(crate) fn set_thinking_enabled(&mut self, enabled: bool) {
-        self.thinking_enabled = enabled;
-    }
-
-    /// Returns a reference to the session tracer, if available.
-    pub(crate) fn session_tracer(&self) -> Option<&telemetry::SessionTracer> {
-        self.client.session_tracer()
-    }
-
-    /// Estimated tokens of the request parts that history compaction cannot
-    /// shrink — the rendered system prompt and the tool definitions this
-    /// client attaches — using the same heuristic as the API preflight.
-    pub(crate) fn fixed_request_overhead_tokens(
-        &self,
-        system_prompt: &runtime::SystemPrompt,
-    ) -> usize {
-        let system = (!system_prompt.is_empty()).then(|| system_prompt.render());
-        let tools = self
-            .enable_tools
-            .then(|| filter_tool_specs(&self.tool_registry, self.allowed_tools.as_ref()));
-        api::estimate_request_overhead_tokens(system.as_deref(), tools.as_deref()) as usize
-    }
-}
-
-/// Map an API failure to a runtime error, preserving the one classification
-/// the runtime acts on: a context-window rejection triggers compaction and a
-/// retry instead of ending the turn.
-fn runtime_error_from_api(session_id: &str, error: &api::ApiError) -> RuntimeError {
-    let message = format_user_visible_api_error(session_id, error);
-    if error.is_context_window_failure() {
-        RuntimeError::context_window_blocked(message)
-    } else {
-        RuntimeError::new(message)
-    }
-}
-
-pub(crate) fn resolve_cli_auth_source() -> Result<AuthSource, Box<dyn std::error::Error>> {
-    Ok(resolve_cli_auth_source_for_cwd()?)
-}
-
-pub(crate) fn resolve_cli_auth_source_for_cwd() -> Result<AuthSource, api::ApiError> {
-    resolve_startup_auth_source(|| Ok(None))
-}
-
-#[async_trait]
-impl ApiClient for AnthropicRuntimeClient {
-    async fn send_compaction(
-        &mut self,
-        model: &str,
-        system_prompt: &str,
-        messages: Vec<runtime::ConversationMessage>,
-        max_tokens: u32,
-    ) -> Result<String, RuntimeError> {
-        let request = MessageRequest {
-            model: model.to_string(),
-            max_tokens,
-            messages: convert_messages(&messages),
-            system: Some(system_prompt.to_string()),
-            tools: None,
-            tool_choice: None,
-            stream: false,
-            reasoning_effort: None,
-            cache_hints: None,
-            ..Default::default()
-        };
-
-        let response = self
-            .client
-            .send_message(&request, None)
-            .await
-            .map_err(|error| {
-                let message = format!(
-                    "compaction API error: {}",
-                    format_user_visible_api_error(&self.session_id, &error)
-                );
-                if error.is_context_window_failure() {
-                    RuntimeError::context_window_blocked(message)
-                } else {
-                    RuntimeError::new(message)
-                }
-            })?;
-
-        // Extract text content from the response
-        let text = response
-            .content
-            .iter()
-            .filter_map(|block| match block {
-                OutputContentBlock::Text { text } => Some(text.as_str()),
-                _ => None,
-            })
-            .collect::<Vec<_>>()
-            .join("");
-
-        if text.is_empty() {
-            return Err(RuntimeError::new(
-                "compaction response contained no text content",
-            ));
-        }
-
-        Ok(text)
-    }
-
-    #[allow(clippy::too_many_lines)]
-    async fn stream(&mut self, request: ApiRequest) -> Result<AssistantEventStream, RuntimeError> {
-        if let Some(progress_reporter) = &self.progress_reporter {
-            progress_reporter.mark_model_phase();
-        }
-        let is_post_tool = request_ends_with_tool_result(&request);
-        let cache_hints = (!request.system_prompt.is_empty()).then(|| CacheHints {
-            system_static: Some(request.system_prompt.static_text()),
-            system_dynamic: Some(request.system_prompt.dynamic_text()),
-            breakpoint_last_message: true,
-        });
-        let trace_id = request.trace_id.as_deref();
-        let message_request = MessageRequest {
-            model: self.model.clone(),
-            max_tokens: max_tokens_for_model(&self.model),
-            messages: convert_messages(&request.messages),
-            system: (!request.system_prompt.is_empty()).then(|| request.system_prompt.render()),
-            tools: self
-                .enable_tools
-                .then(|| filter_tool_specs(&self.tool_registry, self.allowed_tools.as_ref())),
-            tool_choice: self.enable_tools.then_some(ToolChoice::Auto),
-            stream: true,
-            reasoning_effort: self.reasoning_effort.clone(),
-            cache_hints,
-            thinking_enabled: self.thinking_enabled,
-            ..Default::default()
-        };
-
-        // When resuming after tool execution, apply a stall timeout on the
-        // first stream event.  If the model does not respond within the
-        // deadline we drop the stalled connection and re-send the request as
-        // a continuation nudge (one retry only).
-        let max_attempts: usize = if is_post_tool { 2 } else { 1 };
-
-        for attempt in 1..=max_attempts {
-            let result = self
-                .try_start_stream(
-                    &message_request,
-                    is_post_tool && attempt == 1,
-                    is_post_tool,
-                    trace_id,
-                )
-                .await;
-            match result {
-                Ok(stream) => return Ok(stream),
-                Err(error)
-                    if error.to_string().contains("post-tool stall") && attempt < max_attempts =>
-                {
-                    // Stalled after tool completion — nudge the model by
-                    // re-sending the same request.
-                }
-                Err(error) => return Err(error),
-            }
-        }
-
-        Err(RuntimeError::new("post-tool continuation nudge exhausted"))
-    }
-}
-
-/// Internal state for the unfold-based `AssistantEventStream` produced by
-/// [`AnthropicRuntimeClient::try_start_stream`].
-#[allow(clippy::struct_excessive_bools)]
-struct CliStreamState {
-    provider_stream: api::MessageStream,
-    session_id: String,
-    emit_output: bool,
-    progress_reporter: Option<InternalPromptProgressReporter>,
-    spinner: Option<SpinnerRef>,
-    pending_tool: Option<(String, String, String, Option<String>)>,
-    block_has_thinking_summary: bool,
-    markdown_stream: MarkdownStreamState,
-    renderer: TerminalRenderer,
-    glyph_state: ResponseGlyphState,
-    buffer: VecDeque<AssistantEvent>,
-    prefetched_next: Option<Option<ApiStreamEvent>>,
-    saw_stop: bool,
-    has_content: bool,
-    received_any_event: bool,
-    done: bool,
-    /// Clone of the provider client used to extract prompt cache at end of stream.
-    client: ApiProviderClient,
-    /// Non-streaming fallback request (used when streaming yields no events).
-    fallback_request: Option<MessageRequest>,
-    /// Optional channel-backed writer that routes output through the iocraft
-    /// render loop instead of writing directly to stdout.
-    output_writer: Option<OutputSender>,
-}
-
-impl CliStreamState {
-    fn pause_output(&self) {
-        if let Some(s) = &self.spinner {
-            s.pause();
-        }
-    }
-
-    fn resume_output(&self) {
-        if let Some(s) = &self.spinner {
-            s.resume();
-        }
-    }
-
-    fn set_thinking_indicator(&self, on: bool) {
-        if let Some(s) = &self.spinner {
-            s.set_thinking(on);
-        }
-    }
-
-    /// Process a single provider event, converting it into zero or more
-    /// [`AssistantEvent`]s pushed onto `self.buffer`.  All I/O (terminal
-    /// rendering) happens synchronously here so that no `dyn Write` reference
-    /// is held across an `.await` point.
-    #[allow(clippy::too_many_lines)]
-    fn process_provider_event(&mut self, event: ApiStreamEvent) -> Result<(), RuntimeError> {
-        let mut stdout = io::stdout();
-        let mut sink = io::sink();
-        let mut writer_clone = self.output_writer.clone();
-        let out: &mut dyn Write = if self.emit_output {
-            if let Some(ref mut w) = writer_clone {
-                w
-            } else {
-                &mut stdout
-            }
-        } else {
-            &mut sink
-        };
-        match event {
-            ApiStreamEvent::MessageStart(start) => {
-                if !start.message.model.is_empty() {
-                    self.buffer
-                        .push_back(AssistantEvent::Model(start.message.model.clone()));
-                }
-                for block in start.message.content {
-                    push_output_block(
-                        block,
-                        out,
-                        &mut self.buffer,
-                        &mut self.pending_tool,
-                        true,
-                        &mut self.block_has_thinking_summary,
-                        &mut self.glyph_state,
-                    )?;
-                }
-            }
-            ApiStreamEvent::ContentBlockStart(start) => {
-                push_output_block(
-                    start.content_block,
-                    out,
-                    &mut self.buffer,
-                    &mut self.pending_tool,
-                    true,
-                    &mut self.block_has_thinking_summary,
-                    &mut self.glyph_state,
-                )?;
-            }
-            ApiStreamEvent::ContentBlockDelta(delta) => match delta.delta {
-                ContentBlockDelta::TextDelta { text } => {
-                    if !text.is_empty() {
-                        if let Some(s) = &self.spinner {
-                            s.add_response_bytes(text.len() as u32);
-                        }
-                        if let Some(progress_reporter) = &self.progress_reporter {
-                            progress_reporter.mark_text_phase(&text);
-                        }
-                        if let Some(rendered) = self.markdown_stream.push(&self.renderer, &text) {
-                            self.pause_output();
-                            let prefixed = self.glyph_state.apply(&rendered);
-                            write!(out, "{prefixed}")
-                                .and_then(|()| out.flush())
-                                .map_err(|error| RuntimeError::new(error.to_string()))?;
-                        }
-                        self.has_content = true;
-                        self.buffer.push_back(AssistantEvent::TextDelta(text));
-                    }
-                }
-                ContentBlockDelta::InputJsonDelta { partial_json } => {
-                    if let Some(s) = &self.spinner {
-                        s.add_response_bytes(partial_json.len() as u32);
-                    }
-                    if let Some((_, _, input, _)) = &mut self.pending_tool {
-                        input.push_str(&partial_json);
-                    }
-                }
-                ContentBlockDelta::ThinkingDelta { thinking } => {
-                    if let Some(s) = &self.spinner {
-                        s.add_response_bytes(thinking.len() as u32);
-                    }
-                    if !self.block_has_thinking_summary {
-                        // Thinking is not surfaced in the transcript at all;
-                        // the spinner's "Reasoning..." mode is the only cue.
-                        self.block_has_thinking_summary = true;
-                        self.set_thinking_indicator(true);
-                    }
-                    push_thinking_event(&mut self.buffer, thinking, None);
-                }
-                ContentBlockDelta::SignatureDelta { .. } => {}
-            },
-            ApiStreamEvent::ContentBlockStop(_) => {
-                if self.block_has_thinking_summary {
-                    // Coming out of a thinking block — restore the default
-                    // spinner state for whatever comes next.
-                    self.set_thinking_indicator(false);
-                }
-                self.block_has_thinking_summary = false;
-                if let Some(rendered) = self.markdown_stream.flush(&self.renderer) {
-                    let prefixed = self.glyph_state.apply(&rendered);
-                    write!(out, "{prefixed}")
-                        .and_then(|()| out.flush())
-                        .map_err(|error| RuntimeError::new(error.to_string()))?;
-                }
-                if let Some((id, name, input, thought_signature)) = self.pending_tool.take() {
-                    let input = if input.is_empty() {
-                        "{}".to_string()
-                    } else {
-                        input
-                    };
-                    if let Some(progress_reporter) = &self.progress_reporter {
-                        progress_reporter.mark_tool_phase(&name, &input);
-                    }
-                    self.pause_output();
-                    writeln!(out, "\n{}", format_tool_call_start(&name, &input))
-                        .and_then(|()| out.flush())
-                        .map_err(|error| RuntimeError::new(error.to_string()))?;
-                    self.glyph_state.visible_col = 0;
-                    self.resume_output();
-                    self.has_content = true;
-                    self.buffer.push_back(AssistantEvent::ToolUse {
-                        id,
-                        name,
-                        input,
-                        thought_signature,
-                    });
-                }
-            }
-            ApiStreamEvent::MessageDelta(delta) => {
-                self.buffer
-                    .push_back(AssistantEvent::Usage(delta.usage.token_usage()));
-            }
-            ApiStreamEvent::MessageStop(_) => {
-                self.saw_stop = true;
-                if let Some(rendered) = self.markdown_stream.flush(&self.renderer) {
-                    let prefixed = self.glyph_state.apply(&rendered);
-                    write!(out, "{prefixed}")
-                        .and_then(|()| out.flush())
-                        .map_err(|error| RuntimeError::new(error.to_string()))?;
-                }
-                self.buffer.push_back(AssistantEvent::MessageStop);
-            }
-        }
-        Ok(())
-    }
-}
-
-impl AnthropicRuntimeClient {
-    /// Start a streaming response, optionally applying a stall timeout on the
-    /// first event for post-tool continuations.  Returns an incremental stream
-    /// of [`AssistantEvent`]s.  Dropping the stream cancels the underlying
-    /// HTTP request.
-    async fn try_start_stream(
-        &mut self,
-        message_request: &MessageRequest,
-        apply_stall_timeout: bool,
-        is_post_tool: bool,
-        trace_id: Option<&str>,
-    ) -> Result<AssistantEventStream, RuntimeError> {
-        let mut provider_stream = self
-            .client
-            .stream_message(message_request, trace_id)
-            .await
-            .map_err(|error| runtime_error_from_api(&self.session_id, &error))?;
-
-        let prefetched_next = if apply_stall_timeout {
-            match tokio::time::timeout(POST_TOOL_STALL_TIMEOUT, provider_stream.next_event()).await
-            {
-                Ok(inner) => {
-                    match inner.map_err(|error| runtime_error_from_api(&self.session_id, &error))? {
-                        Some(event) => Some(Some(event)),
-                        None => {
-                            return Err(RuntimeError::new(
-                                "post-tool stall: model stream ended before first event",
-                            ));
-                        }
-                    }
-                }
-                Err(_elapsed) => {
-                    return Err(RuntimeError::new(
-                        "post-tool stall: model did not respond within timeout",
-                    ));
-                }
-            }
-        } else {
-            None
-        };
-        let received_any_event = prefetched_next.as_ref().is_some_and(|next| next.is_some());
-
-        let state = CliStreamState {
-            provider_stream,
-            session_id: self.session_id.clone(),
-            emit_output: self.emit_output,
-            progress_reporter: self.progress_reporter.clone(),
-            spinner: self.spinner.clone(),
-            pending_tool: None,
-            block_has_thinking_summary: false,
-            markdown_stream: MarkdownStreamState::default(),
-            renderer: TerminalRenderer::new(),
-            glyph_state: ResponseGlyphState::new(query_terminal_width()),
-            buffer: VecDeque::new(),
-            prefetched_next,
-            saw_stop: false,
-            has_content: false,
-            received_any_event,
-            done: false,
-            client: self.client.clone(),
-            fallback_request: Some(build_non_streaming_fallback_request(
-                message_request,
-                is_post_tool,
-            )),
-            output_writer: self.output_writer.clone(),
-        };
-
-        Ok(Box::pin(futures::stream::try_unfold(
-            state,
-            |mut state| async move {
-                // Yield buffered events first.
-                if let Some(event) = state.buffer.pop_front() {
-                    return Ok(Some((event, state)));
-                }
-                if state.done {
-                    return Ok(None);
-                }
-
-                loop {
-                    let next = if let Some(prefetched_next) = state.prefetched_next.take() {
-                        prefetched_next
-                    } else {
-                        state.provider_stream.next_event().await.map_err(|error| {
-                            RuntimeError::new(format_user_visible_api_error(
-                                &state.session_id,
-                                &error,
-                            ))
-                        })?
-                    };
-
-                    let Some(event) = next else {
-                        // Provider stream ended — emit prompt cache and
-                        // synthetic stop if needed, then signal done.
-                        if let Some(record) = state.client.take_last_prompt_cache_record() {
-                            if let Some(evt) = prompt_cache_record_to_runtime_event(record) {
-                                state.buffer.push_back(AssistantEvent::PromptCache(evt));
-                            }
-                        }
-                        if !state.saw_stop && state.has_content {
-                            state.buffer.push_back(AssistantEvent::MessageStop);
-                        }
-
-                        // If stream produced nothing useful, fall back to
-                        // non-streaming request.
-                        if state.buffer.is_empty() && !state.saw_stop {
-                            if let Some(fallback_request) = state.fallback_request.take() {
-                                let response = state
-                                    .client
-                                    .send_message(&fallback_request, None)
-                                    .await
-                                    .map_err(|error| {
-                                        RuntimeError::new(format_user_visible_api_error(
-                                            &state.session_id,
-                                            &error,
-                                        ))
-                                    })?;
-                                // response_to_events does sync I/O (no await),
-                                // so the dyn Write borrow is safe here.
-                                let mut stdout = io::stdout();
-                                let mut sink = io::sink();
-                                let mut writer_clone = state.output_writer.clone();
-                                let out: &mut dyn Write = if state.emit_output {
-                                    if let Some(ref mut w) = writer_clone {
-                                        w
-                                    } else {
-                                        &mut stdout
-                                    }
-                                } else {
-                                    &mut sink
-                                };
-                                let events = response_to_events(response, out)?;
-                                state.buffer.extend(events);
-                                if let Some(record) = state.client.take_last_prompt_cache_record() {
-                                    if let Some(evt) = prompt_cache_record_to_runtime_event(record)
-                                    {
-                                        state.buffer.push_back(AssistantEvent::PromptCache(evt));
-                                    }
-                                }
-                            }
-                        }
-
-                        state.done = true;
-                        return Ok(state.buffer.pop_front().map(|evt| (evt, state)));
-                    };
-                    state.received_any_event = true;
-
-                    // Process the provider event synchronously (all I/O
-                    // happens inside this call, no dyn Write held across
-                    // await points).
-                    state.process_provider_event(event)?;
-
-                    // If we produced any buffered events, yield the first one.
-                    if let Some(event) = state.buffer.pop_front() {
-                        return Ok(Some((event, state)));
-                    }
-                    // Otherwise loop to read more from the provider.
-                }
-            },
-        )))
-    }
-}
-
-/// Returns `true` when the conversation ends with a tool-result message,
-/// meaning the model is expected to continue after tool execution.
-pub(crate) fn request_ends_with_tool_result(request: &ApiRequest) -> bool {
-    request
-        .messages
-        .last()
-        .is_some_and(|message| message.role == MessageRole::Tool)
-}
-
-const POST_TOOL_FINAL_SYNTHESIS_PROMPT: &str = "The previous tool execution is complete. Send a final ordinary assistant message to the user in the user's language. Do not call tools. Only describe actions explicitly confirmed by the tool results. Only list files whose paths are explicitly present in the tool results. If the tool result only created a draft/helper script, say that the final deliverables have not been generated yet and identify the draft script path.";
-
-pub(crate) fn build_non_streaming_fallback_request(
-    request: &MessageRequest,
-    is_post_tool: bool,
-) -> MessageRequest {
-    let mut fallback = MessageRequest {
-        stream: false,
-        ..request.clone()
-    };
-
-    if is_post_tool {
-        fallback.tools = None;
-        fallback.tool_choice = None;
-        fallback
-            .messages
-            .push(InputMessage::user_text(POST_TOOL_FINAL_SYNTHESIS_PROMPT));
-    }
-
-    fallback
-}
+use runtime::ContentBlock;
 
 pub(crate) fn final_assistant_text(summary: &runtime::TurnSummary) -> String {
     summary
@@ -767,437 +90,4 @@ pub(crate) fn collect_prompt_cache_events(
 
 pub(crate) fn max_tokens_for_model(model: &str) -> u32 {
     api::max_tokens_for_model(model)
-}
-
-pub(crate) fn push_output_block(
-    block: OutputContentBlock,
-    out: &mut (impl Write + ?Sized),
-    events: &mut VecDeque<AssistantEvent>,
-    pending_tool: &mut Option<(String, String, String, Option<String>)>,
-    streaming_tool_input: bool,
-    block_has_thinking_summary: &mut bool,
-    glyph_state: &mut ResponseGlyphState,
-) -> Result<(), RuntimeError> {
-    match block {
-        OutputContentBlock::Text { text } => {
-            if !text.is_empty() {
-                let rendered = TerminalRenderer::new().markdown_to_ansi(&text);
-                let prefixed = glyph_state.apply(&rendered);
-                write!(out, "{prefixed}")
-                    .and_then(|()| out.flush())
-                    .map_err(|error| RuntimeError::new(error.to_string()))?;
-                events.push_back(AssistantEvent::TextDelta(text));
-            }
-        }
-        OutputContentBlock::ToolUse {
-            id,
-            name,
-            input,
-            thought_signature,
-        } => {
-            let initial_input = if streaming_tool_input
-                && input.is_object()
-                && input.as_object().is_some_and(serde_json::Map::is_empty)
-            {
-                String::new()
-            } else {
-                input.to_string()
-            };
-            *pending_tool = Some((id, name, initial_input, thought_signature));
-        }
-        OutputContentBlock::Thinking {
-            thinking,
-            signature,
-        } => {
-            // Thinking is never rendered into the transcript (a complete
-            // block, a streamed opening, or a redacted one alike); the
-            // spinner's "Reasoning..." mode is the only cue.
-            if !thinking.is_empty() {
-                *block_has_thinking_summary = true;
-            }
-            push_thinking_event(events, thinking, signature);
-        }
-        OutputContentBlock::RedactedThinking { .. } => {
-            *block_has_thinking_summary = true;
-            glyph_state.visible_col = 0;
-        }
-    }
-    Ok(())
-}
-
-fn push_thinking_event(
-    events: &mut VecDeque<AssistantEvent>,
-    thinking: String,
-    signature: Option<String>,
-) {
-    events.push_back(AssistantEvent::Thinking {
-        thinking,
-        signature,
-    });
-}
-
-pub(crate) fn response_to_events(
-    response: MessageResponse,
-    out: &mut (impl Write + ?Sized),
-) -> Result<VecDeque<AssistantEvent>, RuntimeError> {
-    let mut events = VecDeque::new();
-    let mut pending_tool = None;
-    let mut glyph_state = ResponseGlyphState::new(query_terminal_width());
-
-    for block in response.content {
-        let mut block_has_thinking_summary = false;
-        push_output_block(
-            block,
-            out,
-            &mut events,
-            &mut pending_tool,
-            false,
-            &mut block_has_thinking_summary,
-            &mut glyph_state,
-        )?;
-        if let Some((id, name, input, thought_signature)) = pending_tool.take() {
-            events.push_back(AssistantEvent::ToolUse {
-                id,
-                name,
-                input,
-                thought_signature,
-            });
-        }
-    }
-
-    events.push_back(AssistantEvent::Usage(response.usage.token_usage()));
-    events.push_back(AssistantEvent::MessageStop);
-    Ok(events)
-}
-
-pub(crate) fn prompt_cache_record_to_runtime_event(
-    record: api::PromptCacheRecord,
-) -> Option<PromptCacheEvent> {
-    let cache_break = record.cache_break?;
-    Some(PromptCacheEvent {
-        unexpected: cache_break.unexpected,
-        reason: cache_break.reason,
-        previous_cache_read_input_tokens: cache_break.previous_cache_read_input_tokens,
-        current_cache_read_input_tokens: cache_break.current_cache_read_input_tokens,
-        token_drop: cache_break.token_drop,
-    })
-}
-
-pub(crate) fn convert_messages(messages: &[ConversationMessage]) -> Vec<InputMessage> {
-    let mut result: Vec<InputMessage> = Vec::with_capacity(messages.len());
-    for message in messages {
-        let role = match message.role {
-            MessageRole::System | MessageRole::User | MessageRole::Tool => "user",
-            MessageRole::Assistant => "assistant",
-        };
-        let content = message
-            .blocks
-            .iter()
-            .filter_map(|block| match block {
-                ContentBlock::Text { text } => Some(InputContentBlock::Text { text: text.clone() }),
-                ContentBlock::Image { data, mime_type } => Some(InputContentBlock::Image {
-                    source: ImageSource {
-                        source_type: "base64".to_string(),
-                        media_type: mime_type.clone(),
-                        data: data.clone(),
-                    },
-                }),
-                ContentBlock::Thinking { .. } => None,
-                ContentBlock::ToolUse {
-                    id,
-                    name,
-                    input,
-                    thought_signature,
-                } => Some(InputContentBlock::ToolUse {
-                    id: id.clone(),
-                    name: name.clone(),
-                    input: serde_json::from_str(input)
-                        .unwrap_or_else(|_| serde_json::json!({ "raw": input })),
-                    thought_signature: thought_signature.clone(),
-                }),
-                ContentBlock::ToolResult {
-                    tool_use_id,
-                    output,
-                    is_error,
-                    ..
-                } => Some(InputContentBlock::ToolResult {
-                    tool_use_id: tool_use_id.clone(),
-                    content: vec![ToolResultContentBlock::Text {
-                        text: output.clone(),
-                    }],
-                    is_error: *is_error,
-                }),
-            })
-            .collect::<Vec<_>>();
-        if content.is_empty() {
-            continue;
-        }
-
-        // Merge consecutive Tool-role messages into the previous user-role
-        // InputMessage. The runtime pushes each tool_result as its own
-        // Tool-role ConversationMessage, but Anthropic requires every
-        // `tool_use` in an assistant turn to have its matching `tool_result`
-        // in the SAME next user message — splitting them across consecutive
-        // user messages triggers `messages.N: tool_use ids` 400 errors.
-        // OpenAI and Gemini conversions iterate per content block and emit
-        // their own per-result messages, so merging here is a no-op for them.
-        if matches!(message.role, MessageRole::Tool) {
-            if let Some(last) = result.last_mut() {
-                if last.role == "user"
-                    && last
-                        .content
-                        .iter()
-                        .all(|block| matches!(block, InputContentBlock::ToolResult { .. }))
-                {
-                    last.content.extend(content);
-                    continue;
-                }
-            }
-        }
-
-        result.push(InputMessage {
-            role: role.to_string(),
-            content,
-        });
-    }
-    result
-}
-
-pub(crate) fn filter_tool_specs(
-    tool_registry: &GlobalToolRegistry,
-    allowed_tools: Option<&AllowedToolSet>,
-) -> Vec<ToolDefinition> {
-    tool_registry.definitions(allowed_tools)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn push_thinking_event_buffers_runtime_thinking() {
-        let mut events = VecDeque::new();
-
-        push_thinking_event(
-            &mut events,
-            "hidden reasoning".to_string(),
-            Some("sig".to_string()),
-        );
-        push_thinking_event(&mut events, String::new(), None);
-
-        assert_eq!(events.len(), 2);
-        assert!(matches!(
-            events.pop_front(),
-            Some(AssistantEvent::Thinking {
-                thinking,
-                signature: Some(signature),
-            }) if thinking == "hidden reasoning" && signature == "sig"
-        ));
-        assert!(matches!(
-            events.pop_front(),
-            Some(AssistantEvent::Thinking {
-                thinking,
-                signature: None,
-            }) if thinking.is_empty()
-        ));
-    }
-
-    #[test]
-    fn response_to_events_preserves_thinking_block() {
-        let response = MessageResponse {
-            id: "msg-1".to_string(),
-            kind: "message".to_string(),
-            role: "assistant".to_string(),
-            content: vec![
-                OutputContentBlock::Thinking {
-                    thinking: "considering".to_string(),
-                    signature: None,
-                },
-                OutputContentBlock::Text {
-                    text: "done".to_string(),
-                },
-            ],
-            model: "model".to_string(),
-            stop_reason: None,
-            stop_sequence: None,
-            usage: Default::default(),
-            request_id: None,
-        };
-        let mut out = io::sink();
-
-        let events = response_to_events(response, &mut out).expect("response should convert");
-
-        assert!(matches!(
-            events.front(),
-            Some(AssistantEvent::Thinking { thinking, .. }) if thinking == "considering"
-        ));
-        assert!(events
-            .iter()
-            .any(|event| matches!(event, AssistantEvent::TextDelta(text) if text == "done")));
-    }
-
-    fn tool_use(id: &str, name: &str) -> ConversationMessage {
-        ConversationMessage {
-            role: MessageRole::Assistant,
-            blocks: vec![ContentBlock::ToolUse {
-                id: id.to_string(),
-                name: name.to_string(),
-                input: "{}".to_string(),
-                thought_signature: None,
-            }],
-            usage: None,
-            model: None,
-        }
-    }
-
-    fn tool_use_multi(ids_and_names: &[(&str, &str)]) -> ConversationMessage {
-        ConversationMessage {
-            role: MessageRole::Assistant,
-            blocks: ids_and_names
-                .iter()
-                .map(|(id, name)| ContentBlock::ToolUse {
-                    id: id.to_string(),
-                    name: name.to_string(),
-                    input: "{}".to_string(),
-                    thought_signature: None,
-                })
-                .collect(),
-            usage: None,
-            model: None,
-        }
-    }
-
-    fn tool_result(id: &str, name: &str, output: &str) -> ConversationMessage {
-        ConversationMessage {
-            role: MessageRole::Tool,
-            blocks: vec![ContentBlock::ToolResult {
-                tool_use_id: id.to_string(),
-                tool_name: name.to_string(),
-                output: output.to_string(),
-                is_error: false,
-            }],
-            usage: None,
-            model: None,
-        }
-    }
-
-    #[test]
-    fn convert_messages_merges_consecutive_tool_results_into_single_user_message() {
-        // Assistant emits two parallel tool_use blocks; runtime appends each
-        // tool_result as its own Tool-role ConversationMessage. Anthropic
-        // requires both tool_results in the same next user message — without
-        // the merge, the second tool_use's id has no matching tool_result in
-        // the immediately-following user turn, triggering a 400 error.
-        let messages = vec![
-            tool_use_multi(&[("call_a", "fn_a"), ("call_b", "fn_b")]),
-            tool_result("call_a", "fn_a", "result_a"),
-            tool_result("call_b", "fn_b", "result_b"),
-        ];
-
-        let converted = convert_messages(&messages);
-
-        assert_eq!(converted.len(), 2, "tool_results must be merged");
-        assert_eq!(converted[0].role, "assistant");
-        assert_eq!(converted[1].role, "user");
-        assert_eq!(converted[1].content.len(), 2);
-        for block in &converted[1].content {
-            assert!(matches!(block, InputContentBlock::ToolResult { .. }));
-        }
-    }
-
-    #[test]
-    fn convert_messages_does_not_merge_tool_result_into_plain_user_text() {
-        // If the previous user message is plain text (not a tool_result
-        // bundle), we must not merge a fresh tool_result into it.
-        let messages = vec![
-            ConversationMessage {
-                role: MessageRole::User,
-                blocks: vec![ContentBlock::Text {
-                    text: "hi".to_string(),
-                }],
-                usage: None,
-                model: None,
-            },
-            tool_use("call_a", "fn_a"),
-            tool_result("call_a", "fn_a", "ok"),
-        ];
-
-        let converted = convert_messages(&messages);
-
-        assert_eq!(converted.len(), 3);
-        assert_eq!(converted[0].role, "user");
-        assert!(matches!(
-            converted[0].content[0],
-            InputContentBlock::Text { .. }
-        ));
-        assert_eq!(converted[2].role, "user");
-        assert!(matches!(
-            converted[2].content[0],
-            InputContentBlock::ToolResult { .. }
-        ));
-    }
-
-    #[test]
-    fn post_tool_non_streaming_fallback_disables_tools_and_requests_final_text() {
-        let request = MessageRequest {
-            model: "gemini-3.5-flash".to_string(),
-            max_tokens: 64000,
-            messages: vec![InputMessage::user_tool_result(
-                "call_a",
-                "created file",
-                false,
-            )],
-            tools: Some(vec![ToolDefinition {
-                name: "write_file".to_string(),
-                description: None,
-                input_schema: serde_json::json!({ "type": "object" }),
-            }]),
-            tool_choice: Some(ToolChoice::Auto),
-            stream: true,
-            ..Default::default()
-        };
-
-        let fallback = build_non_streaming_fallback_request(&request, true);
-
-        assert!(!fallback.stream);
-        assert!(fallback.tools.is_none());
-        assert!(fallback.tool_choice.is_none());
-        assert_eq!(fallback.messages.len(), 2);
-        assert!(matches!(
-            fallback.messages[1].content[0],
-            InputContentBlock::Text { ref text } if text.contains("Do not call tools")
-        ));
-        assert!(matches!(
-            fallback.messages[1].content[0],
-            InputContentBlock::Text { ref text } if text.contains("Only describe actions explicitly confirmed by the tool results")
-        ));
-        assert!(matches!(
-            fallback.messages[1].content[0],
-            InputContentBlock::Text { ref text } if text.contains("final deliverables have not been generated yet")
-        ));
-    }
-
-    #[test]
-    fn regular_non_streaming_fallback_preserves_tool_configuration() {
-        let request = MessageRequest {
-            model: "gpt-5.5".to_string(),
-            max_tokens: 64000,
-            messages: vec![InputMessage::user_text("hello")],
-            tools: Some(vec![ToolDefinition {
-                name: "read_file".to_string(),
-                description: None,
-                input_schema: serde_json::json!({ "type": "object" }),
-            }]),
-            tool_choice: Some(ToolChoice::Auto),
-            stream: true,
-            ..Default::default()
-        };
-
-        let fallback = build_non_streaming_fallback_request(&request, false);
-
-        assert!(!fallback.stream);
-        assert!(fallback.tools.is_some());
-        assert_eq!(fallback.tool_choice, Some(ToolChoice::Auto));
-        assert_eq!(fallback.messages.len(), 1);
-    }
 }
